backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
supabase>=2.5
python-dotenv>=1.0
